HCS-7: back - test return error for delete couple spending category if has spending

+++ app/Http/Livewire/CoupleSpendingCategories/Destroy.php
@@ -13,6 +13,12 @@ class Destroy extends Component
     public function save(): void
     {
 
+        if ($this->category->spendings()->count() > 0) {
+            $this->addError('category', 'This category has spendings');
+
+            return;
+        }
+
         $this->category->delete();
 
         $this->emit('couple-spending-category::deleted');
